Convert Profile to NamedTuple
It's more compact, still has type annotations and does not need linter
pragmas about not having methods and whatnot.

But the main reason for the change is that the class was implementing a
dunder method (to wit, __eq__()) just because a comparison was done in a
test unit. This is utterly wrong. The test unit has been modified, it is
now more explicit and the Profile type, much simpler.