Coalesce Messages and Debug string constants
Since all string constants, in the end, are intended to be read by users
and the only real discrimination is that some of them will be used in a
more detailed log file, and since this distinction can be quite volatile
(a message which NOW only appears in the detailed log file may be needed
in the future in the basic, end-user oriented, log file), it no longer
makes a lot of sense to separate normal messages and debug-like messages
and having all in the same place can in fact be useful.

So, the code has been simplified and both Messages and Debug strings are
now coalesced under Messages.

Test units have been adapted to this change.