Improve winerror handling in excepthook()
Before, winerror was checked so its value was not None, which is weird
but currently it can happen for some OSError derived exceptions. If it
was None, a default marker was used instead to signal the fact that the
real value was useless.

Now, if winerror is ANY 'falsy' value, including 0, the marker is used.

The reason behind this is that if an OSError exception has happened, the
winerror value CANNOT be 0 (which means success). If it is 0, then the
value is equally useless and the marker has to be used, too.