Improve errno handling in excepthook()
Before, errno was checked so its value was not None, which is weird but
currently it can happen for some OSError derived exceptions. If it was
None, a default marker was used instead to signal the fact that the real
value was useless or missing.

This has a potential problem: errno can be missing (so, None) but it can
also be damaged or just out of bounds, and in that case no corresponding
string will exist in error.errorcode mapping.

The only real solution to this is to try to get the corresponding string
message for the current errno value, and if that causes a KeyError or an
IndexError. In that case the default marker is used.