Fix text fixture for unreadable files
Turns out that icacls now removes any grant before applying deny entries
so the test fixture was failing on modern Windows versions because the
permissions for the newly created unreadable file could not be restored.

Now, a different approach is used, where the file will not have any ACL
so it will not be readable by the current user. After that, it can be
removed anyway and the fixture works as before.

Hacky, but it works and it is simple enough for a test fixture.