Fix resolve_file_url() after geturl() changes
Function geturl() used to properly add forward slashes to 'file:' URLs
if it detected the path part started with a drive letter in Windows, or
at least that was what was happening. Not it does not that, and in fact
that is the proper behaviour.

As such, resolve_file_url() now adds a forward slash to absolute path
components so they resolve properly and have the correct forward slashes
as specified by RFC 8089. Please note that the same RFC allows for file:
URLs WITHOUT those slashes, but for backwards compatibility it is better
to support the full slashed syntax.