Replace httpbin.org in test suite
It was used for mocking/testing certain URL errors but it was timing out
lately, and being quite unreliable.

In the future a self-made mock server may be used, but for now the mock
service has been replaced and refactored, so it is easier to replace in
the future.

For now, httpcan.org is used.