Remove version.py and replace versioning scheme
The version string is now managed automatically by the building system,
using the tag markers from git, so the version.py file is superfluous.

Instead of using Semantic Versioning 2.0, as before, the scheme defined
by the Python Packaging Authority, as described in the documentation at
https://packaging.python.org/specifications/version-specifiers/ is now
used. For now only the release segment and optionally (for development
versions) a dev release segment and a local version identifier are used.

For released versions both schemes are compatible, since the string will
be in the form MAJOR.MINOR.PATCH, identical to Semantic Versioning 2.0,
but development releases include a non compatible dev release string.