Remove network accesses in test units
For some tests, httpcan.org was used to simulate network accesses, and
even though it is much more reliable than httpbin.org, in the end it is
a VERY BAD idea to do actual network accesses in the test suite.

So, all network accesses have been replaced with mock operations.

For testing UTF-8 data processing, a new file has been included, instead
of going to a web page to retrieve such data.